fix: More effort to remove circular imports
RainDoh.ts and SpookyPutty.ts both have circular import issues. Find
offending functions and extract them to putty-likes.ts.
