+++ src/lib.ts
@@ -26,8 +26,6 @@ import {
 
 import { $class } from "./template-string";
 import { get } from "./property";
-import { getRainDohBlackBoxCopiesMade } from "./resources/RainDoh";
-import { getSpookyPuttySheetCopiesMade } from "./resources/SpookyPutty";
 
 /**
  * Returns the current maximum Accordion Thief songs the player can have in their head
@@ -309,8 +307,4 @@ export function getFoldGroup(item: Item): Item[] {
 
 export function getZapGroup(item: Item): Item[] {
   return Object.keys(getRelated(item, "zap")).map((i) => Item.get(i));
-}
-
-export function getTotalPuttyLikeCopiesMade(): number {
-  return getSpookyPuttySheetCopiesMade() + getRainDohBlackBoxCopiesMade();
 }
