+++ src/resources/RainDoh.ts
@@ -1,11 +1,6 @@
 import { use } from "kolmafia";
 
-import { Copier } from "../Copier";
-import {
-  getFoldGroup,
-  getTotalPuttyLikeCopiesMade,
-  have as haveItem,
-} from "../lib";
+import { getFoldGroup, have as haveItem } from "../lib";
 import { get } from "../property";
 import { $item } from "../template-string";
 
@@ -19,14 +14,6 @@ export function getRainDohBlackBoxCopiesMade(): number {
   return Math.max(0, get("_raindohCopiesMade"));
 }
 
-export function couldUseRainDohBlackBox(): boolean {
-  return (
-    have() &&
-    getRainDohBlackBoxCopiesMade() < 5 &&
-    getTotalPuttyLikeCopiesMade() < 6
-  );
-}
-
 export function getRainDohBlackBoxMonster(): Monster | null {
   return get("rainDohMonster");
 }
@@ -34,11 +21,3 @@ export function getRainDohBlackBoxMonster(): Monster | null {
 export function useRainDohBlackBox(): boolean {
   return use(box);
 }
-
-export const RainDohBlackBox = new Copier(
-  () => couldUseRainDohBlackBox(),
-  null,
-  () => couldUseRainDohBlackBox(),
-  () => getRainDohBlackBoxMonster(),
-  () => useRainDohBlackBox()
-);
